# pkg/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  auth.Token:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
        type: string
    type: object
...
  models.Message:
    properties:
      message:
        description: A standard message.
        type: string
    type: object
  models.Overview:
    properties:
      projects:
        type: integer
      shares:
        $ref: '#/definitions/models.ShareCounts'
      tasks:
        type: integer
      teams:
        type: integer
      users:
        type: integer
    type: object
...
  models.ShareCounts:
    properties:
      link_shares:
        type: integer
      team_shares:
        type: integer
      user_shares:
        type: integer
    type: object
...
  models.UserGeneralSettings:
    properties:
      default_project_id:
        type: integer
      discoverable_by_email:
        type: boolean
      discoverable_by_name:
        type: boolean
      email_reminders_enabled:
        type: boolean
      extra_settings_links:
        additionalProperties: {}
        description: Server/OpenID-provided; populated on read, ignored on write.
        type: object
      frontend_settings: {}
      language:
        type: string
      name:
        type: string
      overdue_tasks_reminders_enabled:
        type: boolean
      overdue_tasks_reminders_time:
        type: string
      timezone:
        type: string
      week_start:
        maximum: 6
        minimum: 0
        type: integer
    type: object
...
  user.EmailConfirm:
    properties:
      token:
        description: The email confirm token sent via email.
        type: string
    type: object
...
  user.Login:
    properties:
      long_token:
        description: If true, the token returned will be valid a lot longer than default.
          Useful for "remember me" style logins.
        type: boolean
      password:
        description: The password for the user.
        type: string
      totp_passcode:
        description: The totp passcode of a user. Only needs to be provided when enabled.
        type: string
      username:
        description: The username used to log in.
        type: string
    type: object
...
  v1.LinkShareAuth:
    properties:
      password:
        type: string
    type: object
...
  v1.UserPassword:
    properties:
      new_password:
        maxLength: 72
        minLength: 8
        type: string
      old_password:
        type: string
    type: object
...
  v1.UserWithSettings:
    properties:
      auth_provider:
        type: string
      bot_owner_id:
        description: |-
          BotOwnerID is the ID of the owning (human) user if this user is a bot.
          A non-zero value means this user is a bot and cannot authenticate via password.
        type: integer
      created:
        description: A timestamp when this task was created. You cannot change this
          value.
        type: string
      deletion_scheduled_at:
        type: string
      email:
        description: The user's email address.
        maxLength: 250
        type: string
      id:
        description: The unique, numeric id of this user.
        type: integer
      is_admin:
        type: boolean
      is_local_user:
        type: boolean
      name:
        description: The full name of the user.
        type: string
      pending_email:
        type: string
      settings:
        $ref: '#/definitions/models.UserGeneralSettings'
      updated:
        description: A timestamp when this task was last updated. You cannot change
          this value.
        type: string
      username:
        description: The username of the user. Is always unique.
        maxLength: 250
        minLength: 1
        type: string
    type: object
  web.HTTPError:
    properties:
      code:
        type: integer
      i18n_params:
        additionalProperties:
          type: string
        description: I18nParams carries Message's dynamic values, keyed by the client's
          translation placeholder names, so clients can localise the error.
        type: object
      message:
        type: string
    type: object
info:
  contact:
    name: Fenster
    url: https://github.com/MBeggiato/fenster
  description: |-
    # Pagination
    Every endpoint capable of pagination will return two headers:
    * `x-pagination-total-pages`: The total number of available pages for this request
    * `x-pagination-result-count`: The number of items returned for this request.
    # Permissions
    All endpoints which return a single item (project, task, etc.) - no array - will also return a `x-max-permission` header with the max permission the user has on this item as an int where `0` is `Read Only`, `1` is `Read & Write` and `2` is `Admin`.
    This can be used to show or hide ui elements based on the permissions the user has.
    # Errors
    All errors have an error code and a human-readable error message in addition to the http status code. You should always check for the status code in the response, not only the http status code.
    Due to limitations in the swagger library we're using for this document, only one error per http status code is documented here. Make sure to check the [error docs](https://vikunja.io/docs/errors/) in the upstream Vikunja documentation for a full list of available error codes.
    # Authorization
    **JWT-Auth:** Main authorization method, used for most of the requests. Needs `Authorization: Bearer <jwt-token>`-header to authenticate successfully.

...

    **BasicAuth:** Only used when requesting tasks via CalDAV.
...
  license:
    name: AGPL-3.0-or-later
...
paths:
...
  /admin/overview:
    get:
      description: Returns per-instance counts (users, projects, shares) plus version
        and license info. Instance-admin only, gated by the admin_panel feature.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Overview'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.HTTPError'
      security:
      - JWTKeyAuth: []
      summary: Admin overview
      tags:
      - admin
...
  /login:
    post:
      consumes:
      - application/json
      description: Logs a user in. Returns a JWT-Token to authenticate further requests.
      parameters:
      - description: The login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/user.Login'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.Token'
        "400":
          description: Invalid user password model.
          schema:
            $ref: '#/definitions/models.Message'
        "403":
          description: Invalid username or password.
          schema:
            $ref: '#/definitions/models.Message'
        "412":
          description: Invalid totp passcode.
          schema:
            $ref: '#/definitions/models.Message'
      summary: Login
      tags:
      - auth
...
  /shares/{share}/auth:
    post:
      consumes:
      - application/json
      description: Get a jwt auth token for a shared project from a share hash.
      parameters:
      - description: The password for link shares which require one.
        in: body
        name: password
        required: true
        schema:
          $ref: '#/definitions/v1.LinkShareAuth'
      - description: The share hash
        in: path
        name: share
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: The valid jwt auth token.
          schema:
            $ref: '#/definitions/auth.Token'
        "400":
          description: Invalid link share object provided.
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/models.Message'
      summary: Get an auth token for a share
      tags:
      - sharing
...
  /user:
    get:
      consumes:
      - application/json
      description: Returns the current user object with their settings.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.UserWithSettings'
        "404":
          description: User does not exist.
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal server error.
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - JWTKeyAuth: []
      summary: Get user information
      tags:
      - user
  /user/confirm:
    post:
      consumes:
      - application/json
      description: Confirms the email of a newly registered user.
      parameters:
      - description: The token.
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/user.EmailConfirm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Message'
        "412":
          description: Bad token provided.
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/models.Message'
      summary: Confirm the email of a new user
      tags:
      - user
...
  /user/password:
    post:
      consumes:
      - application/json
      description: Lets the current user change its password.
      parameters:
      - description: The current and new password.
        in: body
        name: userPassword
        required: true
        schema:
          $ref: '#/definitions/v1.UserPassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Message'
        "400":
          description: Something's invalid.
          schema:
            $ref: '#/definitions/web.HTTPError'
        "404":
          description: User does not exist.
          schema:
            $ref: '#/definitions/web.HTTPError'
        "500":
          description: Internal server error.
          schema:
            $ref: '#/definitions/models.Message'
      security:
      - JWTKeyAuth: []
      summary: Change password
      tags:
      - user
...
  /user/token:
    post:
      consumes:
      - application/json
      description: Returns a new valid jwt link share token. Only works for link share
        tokens.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.Token'
        "400":
          description: Only link share tokens can be renewed.
          schema:
            $ref: '#/definitions/models.Message'
      summary: Renew link share token
      tags:
      - auth
  /user/token/refresh:
    post:
      description: Exchanges the refresh token cookie for a new short-lived JWT.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.Token'
        "401":
          description: Invalid or expired refresh token.
          schema:
            $ref: '#/definitions/models.Message'
      summary: Refresh user token
      tags:
      - auth
...
securityDefinitions:
  BasicAuth:
    type: basic
  JWTKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"